requirements.txt
streamlit
pandas
streamlit_lottie
mysql-connector-python